OvmfPkg: S3 Suspend: use SMM instances for LockBoxLib library class
LockBoxLib has two sets of implementations. The Null implementation (which
is the OVMF default) is not useful for us any longer; we need the library
instances that back the LockBox structure with SMRAM -- the SmmLockBoxLib
variants.

These variants (library instances) are for:
- DXE_DRIVER, DXE_RUNTIME_DRIVER, UEFI_DRIVER, UEFI_APPLICATION: usable
  from "untrusted", 3rd party drivers; communicates with the SmmLockBox
  driver (which is trusted and runs in SMM).
- PEIM: same purpose but for PEIMs.
- DXE_SMM_DRIVER: this is the low level driver that runs inside the
  trusted SmmLockBox driver, and manipulates SMRAM.

  LockBoxLib [SmmLockBoxDxeLib]
    EFI_SMM_COMMUNICATION_PROTOCOL [PiSmmCore]
      EFI_SMM_ACCESS2_PROTOCOL [EmuSmmDxe]
      EFI_SMM_CONTROL2_PROTOCOL [EmuSmmDxe]
    gEfiSmmLockBoxCommunicationGuid [SmmLockBox]
      EFI_SMM_ACCESS2_PROTOCOL [EmuSmmDxe]
      gEfiSmmSwDispatch2ProtocolGuid [unneeded / EmuSmmDxe]
      LockBoxLib [SmmLockBoxSmmLib]

In this patch, we resolve the library class only for DXE_RUNTIME_DRIVER
and DXE_SMM_DRIVER (which is sufficient to build and run OvmfPkg). We'll
add the other resolutions when we need them.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>