UefiCpuPkg: S3Resume2Pei: allocate return stacks explicitly
S3RestoreConfig2() can optionally switch to the SMM S3 Resume Entry Point
and ask it to transfer to S3ResumeExecuteBootScript().

Similarly, S3ResumeExecuteBootScript() switches explicitly to the boot
script executor, and asks it to transfer to S3ResumeBootOs().

Currently the stack pointers specified for the SMM S3 Resume Entry Point
and the boot script executor to use after returning are derived from
addresses of the first local variables in S3RestoreConfig2() and
S3ResumeExecuteBootScript(), respectively.

Unfortunately, compilers are free to juggle local variables as they
please, including alignment and relation to the stack pointer. Replace the
current derivation of ReturnStackPointer with a new function that
explicitly allocates correctly aligned memory for stacks.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>