MdeModulePkg: DxeLoad: UpdateStackHob(): honor original allocation type
The unique Stack HOB describes the initial memory area used as stack and
PEI heap.

UpdateStackHob() -- called by the various platforms' HandOffToDxeCore()
functions -- modifies this HOB before it is passed to the DXE phase. The
UpdateStackHob() function repoints the Stack HOB at the memory area that
the DXE phase will actually use as stack.

However, as the function explains, the original area must be kept around
as well (DXE must not reclaim it), because  "some PEIMs may [...] keep key
information on the stack". For this end, UpdateStackHob() creates a new
memory allocation HOB that covers the original area.

Currently, the new memory allocation HOB doesn't honor / mirror the
original memory allocation type; it describes the old area invariably as
EfiBootServicesData. Normally this is alright, but it causes problems if
the original stack area was destined for a lifetime longer than
EfiBootServicesData, for example EfiACPIMemoryNVS.

Change UpdateStackHob() so that it preserves not only the boundaries of
the original area, but its allocation type too, in the new HOB.

The patch makes no visible difference if the original allocation type too
was EfiBootServicesData.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>