OvmfPkg: S3 Resume: preallocate PEI_S3_RESUME_STATE for S3Resume2Pei
We don't need this reservation / preallocation in order to save data from
suspend to resume. (The ZeroMem() is added to make this obvious.) It
rather acts as a placeholder after cold boot, to keep the OS away.

S3Resume2Pei needs a PEI_S3_RESUME_STATE object. The object cannot be on
the stack (because it is used across stack switching). It also can't be a
global variable (because S3Resume2Pei in general could run from flash).
S3Resume2Pei cannot allocate it dynamically either (because that could
trample OS data). Hence we must preallocate the object for S3Resume2Pei,
in a spot that the OS never touches.

The lower-level PeiServicesAllocatePages() function
[MdePkg/Library/PeiServicesLib] is called rather than the higher-level
AllocatePool() / AllocatePages() functions
[MdePkg/Library/PeiMemoryAllocationLib] because the latter wrappers don't
provide access to EfiACPIMemoryNVS.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>