OvmfPkg: introduce EmuNvramLib
In the following patches we'll need three kinds of storage that are
reserved from the OS and survive S3 suspend. (Non-volatility across cold
reboot is not needed.) This trivial library ensures that all parties will
have the same notion about where in guest RAM the various sections of the
NVRAM are stored.

Placement and size of the emulated NVRAM are arbitrary, we only take care
that it is above the decompressed firmware image, and safely apart from
the temporary output and scratch buffers.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>