OvmfPkg: AcpiS3SaveDxe: depend on SmmLockBox explicitly
AcpiS3SaveDxe has a runtime dependency (enforced via ASSERT ()) on
SmmLockBox (which in turn depends on the SMRAM we emulate as part of the
NVRAM). Make the dependency explicit in the Depex section of the driver,
similarly to what BootScriptExecutorDxe does. When the SMRAM is not
available, not loading the driver is preferable to triggering an ASSERT()
in it.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>