MdePkg: introduce RbTreeLib
edk2 should have a fast and easy-to-use associative array (a dictionary)
type.

Red-black trees have O(log(n)) worst case time complexity for lookup,
insertion, and deletion (where n is the number of nodes in the tree). They
support minimum and maximum lookup with the same time complexity, hence
red-black trees double as priority queues as well.

Given an iterator to a red-black tree node, getting the next or previous
node (which corresponds to the ordered successor or the predecessor,
respectively, according to the user-defined ordering) is O(log(n)) as
well.

This patch introduces the RbTreeLib library class, and adds a BASE
instance for it. Memory allocation of nodes is delegated to
MemoryAllocationLib.

Currently all nodes are allocated with AllocatePool(), in Boot Services
Data type memory.

The RbTreeLib interface operates with four concepts:
- tree,
- node,
- user structure,
- user key.

User structures do not embed tree nodes; tree nodes reference user
structures. This allows a user structure to be linked into any number of
trees at runtime. In addition, the reference from a tree node to a user
structure is the "navigation" kind, rather than the "ownership" kind per
default; if any reference counting of user structures is needed, that
should be implemented inside the user structures.

Read-write operations (insertion, deletion) on a tree must be serialized
by the client programmer against all other operations on the same tree, in
the multi-threading / timer callback sense.

(The above is all described in code comments as well.)

The library class approaches a tree with an emphasis on iterators, ie.
pointers to tree nodes. The API description, and the example program in
the next patch, explain and demonstrate why this is a flexible design. In
particular, the library allows several simultaneous, overlapping
iterations through the same tree, with intervening insertions and
deletions. An existent iterator is never invalidated by any operation
except its specific deletion.

The code reflects the Binary Search Trees and Red-Black Trees chapters of
Introduction to Algorithms, by Cormen, Leiserson, Rivest. One point where
the implementation diverges is the first phase of the RbTreeDelete()
operation. During that phase, the book's algorithm copies the key and
other business *contents* of the successor node (in case the successor
node is affected), and releases the successor node (instead of the node
that the user requested to delete).

While semantically correct, this would break the above iterator validity
guarantee. This implementation replaces the copying of business contents
between nodes with suitable relinking of nodes, so that all iterators
(except the one whose deletion is being requested) remain valid.

I had written this code originally in approx. 2000. I personally own the
copyright of that version and am hereby relicensing it to Red Hat, under
the BSDL. I had used the original code in a few personal projects since,
for example in the lbzip2-0.x parallel (de)compressor, and now I've ported
the library to edk2. Both during the original implementation and now
during the porting I verified all the cases and their proofs as rigorously
as I could, on paper. (NB, I couldn't find any errors in the 2000 code
now.)

During the porting to edk2, I documented all those cases in code comments
as well (at least half of the source is documentation). These comments are
not blind copies of diagrams from the Algorithms book, nor are they copies
from my original code -- I've done them all fresh now, and I've only
matched the results against the book. Reviewers are invited to sit down
with a pen, some paper, the book, and the code.

The library provides an RbTreeValidate() function that verifies the
internal red-black properties of the tree. This function helps with unit
testing, and is not linked into binaries that don't invoke it.

A note about diagrams: edges represented by backslash (\) characters are
often written as "\_", ie. with a following underscore. This is because
line-trailing backslashes are processed very early in compilation (in
translation phase 2), "splicing physical source lines to form logical
source lines". Since the edk2 coding style requires "//" comments for such
documentation, a trailing backslash would splice the next physical line
into the "scope" of the comment. To prevent this, trailing backslashes are
defanged by appending underscores, which should be visually bearable.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>