AppPkg: introduce RbTreeTest
In this patch a small application is added to AppPkg, with the following
two goals:
- demonstrate how to use RbTreeLib,
- allow users to test and "fuzz" RbTreeLib, entering RbTreeLib API
  "commands" interactively, or providing them from a script file.

A shell script is also provided that generates such an RbTreeTest command
script.

RbTreeTest validates the internal red-black properties of the tree by
calling RbTreeLib's RbTreeValidate() function after each read-write
operation.

The RbTreeTest application's debugging environment is strictly specified
in the DSC file, because RbTreeTest is entirely useless for unit testing
without full ASSERT() enablement.

The RbTreeTest application deliberately doesn't follow the edk2 coding
style in the following:
- const vs. CONST,
- void vs. VOID,
- assert() vs. ASSERT(),
- calloc() and free() vs. AllocateZeroPool() and FreePool(),
- integer types.

This is because RbTreeTest is a standard C application, not a UEFI
application per se. In particular it relies on stdio. INTN, EFIAPI and
CONST VOID are used only in two places, where we provide the comparator
callbacks to RbTreeLib. Proper range checking is ensured for integers.

The application takes command input from stdin or a file (if the user
requests it), sends command output to stdout or a file (if the user
requests it), prints debug output to the console (as other AppPkg
applications do when debugging is enabled for them), and prints
diagnostics to stderr (like well behaved standard C programs should).

Input/output selection is implemented manually because the old shell
doesn't support input redirection at all, and because the new shell's
input redirection does not co-operate with fgets() for the time being.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>