OvmfPkg: AcpiPlatformDxe: implement QEMU's full ACPI table loader interface
Recent changes in the QEMU ACPI table generator have shown that our
limited client for that interface is insufficient and/or brittle.

Implement the full interface utilizing OrderedCollectionLib for addressing
fw_cfg blobs by name.

In addition, we stop using EFI_ACPI_TABLE_PROTOCOL in the QEMU loader
client, because:
- splitting the blobs for EFI_ACPI_TABLE_PROTOCOL is just untenable short
  of a complete ACPI parser,
- and it is fully sufficient to install the RSD PTR as an EFI
  configuration table for the guest OS to see everything.

In short, for this interface, EFI_ACPI_TABLE_PROTOCOL is an unwanted and
restrictive convenience; let's stop using it.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>