OvmfPkg: AcpiPlatformDxe: remove current ACPI table loader
In the next patch we rewrite the client code for QEMU's fw_cfg ACPI table
loader interface. In order to avoid randomly intermixed hunks in that
patch, first remove the old code cleanly.

We remove the InstallQemuLinkedTables() function and empty the
InstallAllQemuLinkedTables() function. We keep CheckRsdp(), and the way we
call it ensures that

- compilers won't complain about it being a unused static function,

- InstallAllQemuLinkedTables() will return constant EFI_NOT_FOUND to
  AcpiPlatformEntryPoint(), causing the latter to proceed to OVMF's
  builtin tables.

This way the history remains bisectable and the new client gets a clean
start in the next patch.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>