OvmfPkg: QemuBootOrderLib: expose QEMU's "-boot menu=on,splash-time=N"
The command line option

  -boot menu=on,splash-time=N

of QEMU causes SeaBIOS to wait N milliseconds for the user to enter the
boot menu. During the wait, the user can opt to enter the boot menu, or
interrupt the wait and proceed to booting at once. If the wait interval
elapses, SeaBIOS boots as it normally would.

We can approximate this behavior quite well for edk2's virtual platforms
because the Intel BDS front page already supports a progress bar, with
semantics similar to the above. Let's distill the fw_cfg bits underlying
"-boot menu=on,splash-time=N" for the BDS policies, in the form of a
timeout value they can pass to Intel's PlatformBdsEnterFrontPage().

If the boot menu is not requested, we return
"gEfiIntelFrameworkModulePkgTokenSpaceGuid.PcdPlatformBootTimeOut", which
is what the virtual platforms use right now.

RHBZ: https://bugzilla.redhat.com/show_bug.cgi?id=1170507

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>