OvmfPkg/AcpiPlatformDxe: dynamic dependency on PCI enumeration completion
The depexes in AcpiPlatformDxe are not flexible enough for client
platforms to express dynamically whether or not AcpiPlatformDxe should
wait with ACPI table download and installation until PCI enumeration
completes.

AcpiPlatformDxe shall wait for PCI enumeration completion as a default,
but we introduce a new, empty protocol with which the platform can signal
AcpiPlatformDxe to proceed at once. Each platform can install this
protocol (with a NULL interface on a new handle) whenever it sees fit.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>