OvmfPkg: ArmXenHypercallLib: add empty constructor
In the next patch we'll add a simple query function to the XenHypercallLib
library class that is supposed to be called by initialization code in
modules. Among those, in constructors of dependent libraries too.

Library construction ordering is ensured only between libraries with
constructors, plus we shouldn't allow a dependent library with a
constructor to call into ArmXenHypercallLib (the simple query function)
before ArmXenHypercallLib is constructed itself. For this reason,
introduce an empty constructor.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>