MdeModulePkg: SmmIplEntry(): don't suppress SMM core startup failure
When the ExecuteSmmCoreFromSmram() function fails, SmmIplEntry() restores
the SMRAM range to EFI_MEMORY_UC. However, it saves the return value of
gDS->SetMemorySpaceAttributes() in the same Status variable that contains
the return value of ExecuteSmmCoreFromSmram(). Therefore, if
gDS->SetMemorySpaceAttributes() succeeds, the failure of
ExecuteSmmCoreFromSmram() is masked, and Bad Things Happen (TM).

Introduce a temporary variable just for the return value of
gDS->SetMemorySpaceAttributes().

Cc: Feng Tian <feng.tian@intel.com>
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>