OvmfPkg: build PiSmmCpuDxeSmm for -D SMM_REQUIRE
At this point we can enable building PiSmmCpuDxeSmm.

The SmmLib dependency is resolved to MdePkg's SmmLibNull instance. First,
this matches the resolution in "UefiCpuPkg/UefiCpuPkg.dsc". Second, from
this library class, PiSmmCpuDxeSmm calls the ClearSmi() interface only,
and that API is a no-op on Q35 anyway.

CPU specific features, like SMRR detection, and functions that are used to
initialize SMM and process SMIs, are abstracted through the
SmmCpuFeaturesLib class for the PiSmmCpuDxeSmm module. Resolve it to the
one implementation under UefiCpuPkg.

SmmCpuPlatformHookLib provides platform specific functions that are used
to initialize SMM and process SMIs. Resolve it to the one Null instance
provided by UefiCpuPkg, which is expected to work for most platforms.

PiSmmCpuDxeSmm is not yet intended to function correctly (especially with
regard to S3); the upcoming, trimmed down CpuMpDxe port (under the name
CpuS3DataDxe) will be necessary for that.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>