OvmfPkg: import CpuConfigLib from Quark_EDKII_v1.1.0/IA32FamilyCpuBasePkg
The "SocketLga775Lib.h" file is an internal dependency of the library
instance being imported. It should belong under Include/IndustryStandard,
and not be listed in the DEC file (since it is not a standalone library
class), but let's follow the structure of
Quark_EDKII_v1.1.0/IA32FamilyCpuBasePkg.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>