UefiCpuPkg: CpuDxe: optionally save MTRR settings to AcpiNVS memory block
The

  Quark_EDKII_v1.1.0/IA32FamilyCpuBasePkg/CpuArchDxe

driver provides the following capability in its implementation of
EFI_CPU_ARCH_PROTOCOL: whenever the SetMemoryAttributes() member is used
(directly, or indirectly via gDS->SetMemorySpaceAttributes()) to change
MTRR settings, the complete set of settings is immediately reflected to an
AcpiNVS memory block.

(The address of said block is published via a dynamic PCD.)

This feature is necessary for supporting SMM-capable S3 resume, hence port
it to UefiCpuPkg. In IA32FamilyCpuBasePkg the feature is always built in;
here we make it controllable with a feature PCD.

Cc: Jeff Fan <jeff.fan@intel.com>
Cc: Chen Fan <chen.fan.fnst@cn.fujitsu.com>
Cc: Jordan Justen <jordan.l.justen@intel.com>
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>