UefiCpuPkg: CpuDxe: broadcast MTRR changes to APs
The

  Quark_EDKII_v1.1.0/IA32FamilyCpuBasePkg/CpuArchDxe

driver applies any MTRR changes to APs, if the EFI_MP_SERVICES_PROTOCOL is
available. We should do the same; port it to UefiCpuPkg.

The broadcast should occur even if PcdCpuSyncMtrrToAcpiNvs is set to
FALSE; in that case we just use a local MTRR_SETTINGS buffer, rather than
the firmware-wide "mMtrrTable" object.

Cc: Jeff Fan <jeff.fan@intel.com>
Cc: Chen Fan <chen.fan.fnst@cn.fujitsu.com>
Cc: Jordan Justen <jordan.l.justen@intel.com>
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>