OvmfPkg: PiSmmCpuDxeSmm: eliminate CpuConfigLib linkage dependency
The only dependency PiSmmCpuDxeSmm actually has on CpuConfigLib is some
type definitions from the library interface; that is, the header file. No
functions are called and no extern variables are used from the library
instance.

By not linking against CpuConfigLib in our PiSmmCpuDxeSmm port, we can
also avoid importing the dynamic PCD "PcdCpuConfigContextBuffer" to
OvmfPkg later on. That PCD is internal to the CpuConfigLib instance, and
to such a portion of the

  Quark_EDKII_v1.1.0/IA32FamilyCpuBasePkg/CpuMpDxe

driver that we are not going to rely on for bringing SMM to OVMF.

Suggested-by: Paolo Bonzini <pbonzini@redhat.com>
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>