OvmfPkg: import SocketLga775Lib header from Quark_EDKII_v1.1.0/IA32FamilyCpuBasePkg
This header file is another dependency of PiSmmCpuDxeSmm.

It should actually belong under Include/IndustryStandard, and not be
listed in the DEC file (since it is not a standalone library class and
it's not associated with any library instance), but let's follow the
structure of "Quark_EDKII_v1.1.0/IA32FamilyCpuBasePkg" here too.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>