OvmfPkg: QuarkPort/CpuS3DataDxe: handle register tables in ACPI_CPU_DATA
Each one of the "PreSmmInitRegisterTable" and "RegisterTable" fields in
ACPI_CPU_DATA is a pointer to an array of CPU_REGISTER_TABLE objects; one
object per logical processor. Each table carries a variable number of
CPU_REGISTER_TABLE_ENTRY objects; each entry prescribes a specific
register setting for the CPU whose CPU_REGISTER_TABLE the entry is in.

Such entries are added to the tables with the following CpuConfigLib
functions:
- WriteRegisterTableEx(): internal helper,
- WriteRegisterTable(): public function for RegisterTable,
- WritePreSmmInitRegisterTable(): public function for
  PreSmmInitRegisterTable.

* The driver

    Quark_EDKII_v1.1.0/IA32FamilyCpuBasePkg/CpuMpDxe/

  allocates the array of CPU_REGISTER_TABLEs in AcpiNVS for
  ACPI_CPU_DATA.PreSmmInitRegisterTable, and sets the InitialApicID field
  of each table object. (This is then later used by PiSmmCpuDxeSmm to
  match a table against a logical CPU.)

  However, CpuMpDxe never actually adds any register entries to any
  processor's table in the PreSmmInitRegisterTable, either with
  WritePreSmmInitRegisterTable(), or manually.

* CpuMpDxe is theoretically capable of adding register entries (MSR, CRn
  changes etc) to the elements of the array pointed-to by
  ACPI_CPU_DATA.RegisterTable. See the following call tree in CpuMpDxe:

  ProduceRegisterTable()
    MaxCpuidLimitReg()
      WriteRegisterTable(
        ..., MSR_IA32_MISC_ENABLE, N_MSR_LIMIT_CPUID_MAXVAL, ...
        )
    XdReg()
      WriteRegisterTable(
        ..., MSR_IA32_MISC_ENABLE, N_MSR_XD_BIT_DISABLE, ...
        )

  However, both MaxCpuidLimitReg() and XdReg() are gated by feature PCDs;
  we'll just consider those disabled.

Therefore, we'll explicitly instruct PiSmmCpuDxeSmm to do nothing in
response to these tables: we allocate both arrays, but leave them
zero-filled.

That will short-circuit the processing in PiSmmCpuDxeSmm for most logical
CPUs, in the

  (RegisterTableList[Index].InitialApicId == InitApicId)

checks. And when those checks don't fail, and SetProcessorRegister() gets
called, then the

  RegisterTable->TableLength == 0

equality will short-circuit that function.

Suggested-by: Michael Kinney <michael.d.kinney@intel.com>
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>