OvmfPkg: double the SMRAM (TSEG) size for the 64-bit DXE phase builds
When building OVMF with -D SMM_REQUIRE -D SECURE_BOOT_ENABLE, using
gcc-4.8, for the DEBUG build target, and with DEBUG_VERBOSE enabled in
PcdDebugPrintErrorLevel, the build report files report the following
binary sizes for the SMM_CORE and SMM modules, in kilobytes:

  Driver                      Ia32      X64
  ------------------------  ------  -------
  PiSmmCore                  56.75    98.22
  CpuIo2Smm                  13.47    22.72
  SmmLockBox                 24.62    39.69
  PiSmmCpuDxeSmm             66.06   123.03
  FvbServicesSmm             22.22    37.06
  SmmFaultTolerantWriteDxe   35.19    60.06
  VariableSmm               356.88   640.28
  ------------------------  ------  -------
  Total                     575.19  1021.06

The 1 MB default for PcdQ35TsegMbytes, from OvmfPkg.dec, is insufficient
for loading the privileged half of the variable driver (VariableSmm) on
X64. This results in many warnings about the dependent drivers being
impossible to dispatch, and ultimately a DXE core assertion failure.

PcdQ35TsegMbytes can assume the values 1, 2 or 8; set it to 2 for the DSC
files with 64-bit DXE phases. For consistency between the DSC files, spell
out 1 in OvmfPkgIa32.dsc too.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>