UefiCpuPkg/CpuDxe: Prepare final HALT loop for APs
Our current ExitBootServices() callback puts the APs in wait-for-SIPI
state, by sending them an INIT IPI. This works fine as long as the APs are
not woken between ExitBootServices() and the time the runtime OS boots the
APs.

However, if a platform includes PiSmmCpuDxeSmm, and bases the runtime UEFI
services (like the variable services) on SMM and SMIs, then two problem
scenarios can be experienced:

(a) On QEMU, the APs can be (spuriously?) woken by SMIs, between
    ExitBootServices() and the runtime OS's AP boot; for example by
    variable service calls. In this case the APs have no code to continue
    executing after the RSM instruction, which causes them to re-set.

(b) On physical platforms, the APs might not wake from the wait-for-SIPI
    state at all, causing performance (or functional) problems with
    PiSmmCpuDxeSmm, dependent on the BSP/AP sync mode configured therein.

Instead, reboot all APs with a minimal real-mode stub when
ExitBootServices() is called. Each AP shall bump a shared counter
atomically at first, then spiral into an infinite HLT loop. (This occurs
regardless of what each AP has been doing at that moment.) APs can wake
from the HLT due to an SMI, and easily return to the loop upon RSM as
well. Finally, the BSP will be able to ensure in the ExitBootServices()
callback that all APs have moved on to the stub.

This patch introduces the CPU_AP_HLT_LOOP data structure only, with the
related accessor functions. The ExitBootServices() callback is not
modified yet.

Cc: Chen Fan <chen.fan.fnst@cn.fujitsu.com>
Cc: Jeff Fan <jeff.fan@intel.com>
Cc: Jordan Justen <jordan.l.justen@intel.com>
Cc: Michael Kinney <michael.d.kinney@intel.com>
Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Laszlo Ersek <lersek@redhat.com>